Split logic into different files
