+++ Food/FoodApp.swift
@@ -6,20 +6,6 @@
 //
 
 import SwiftUI
-import GoogleSignIn
-
-class AppDelegate: NSObject, UIApplicationDelegate {
-
-    func application(
-        _ application: UIApplication,
-        didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]? = nil
-    ) -> Bool {
-        // Initialization sign-in
-        GIDSignIn.sharedInstance().clientID = "TODO"
-        return true
-    }
-
-}
 
 @main
 struct FoodApp: App {
@@ -29,6 +15,7 @@ struct FoodApp: App {
     var body: some Scene {
         WindowGroup {
             ContentView()
+                .environmentObject(GoogleSignInManager())
         }
     }
